+++ optd-dsl/src/lib.rs
@@ -3,4 +3,5 @@ pub mod gen;
 // mod interpreter;
 pub mod irs;
 pub mod lexer;
-// mod parser;
+pub mod parser;
+pub mod errors;
